backend/requirement.txt
aiohttp==3.13.2
apache-airflow==3.1.1
asyncpg==0.30.0
docling==2.59.0
docling-core==2.50.0
fastapi==0.117.1
fastembed==0.7.3
loguru==0.7.3
neo4j==6.0.2
openai==2.6.1
openai-agents==0.4.2
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pypdfium2==4.30.0
python-dotenv==1.2.1
qdrant-client==1.15.1
sentence-transformers==5.1.2
SQLAlchemy==2.0.44
tenacity==9.1.2
torch==2.9.0
torchvision==0.24.0
transformers==4.57.1
uvicorn==0.38.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0
faker>=22.0.0
factory-boy>=3.3.0